Rename existant CustomField to avoid CustomFieldDefinition
