+++ src/Dtos/CustomField/CustomFieldDto.php
@@ -8,14 +8,30 @@ use OrigamiMp\OrigamiApiSdk\Dtos\ApiResponseDto;
 use OrigamiMp\OrigamiApiSdk\Enums\Dtos\CustomField\CustomFieldDtoTypeEnum;
 use OrigamiMp\OrigamiApiSdk\Exceptions\Dtos\ApiResponseDtoNotConstructableException;
 use OrigamiMp\OrigamiApiSdk\Exceptions\Dtos\CustomField\CustomFieldDtoNotConstructableException;
+use OrigamiMp\OrigamiApiSdk\Traits\Dtos\HasTimestamps;
 
+/**
+ * Represents a Custom Field definition, as returned by the `custom_fields` catalog
+ * endpoint (e.g. `GET custom_fields?filter[model_type]=seller`).
+ *
+ * Distinct from CustomFieldModelValueDto, which represents a value already set on
+ * an entity (Seller, User, ...) and comes from a different, nested JSON shape.
+ */
 class CustomFieldDto extends ApiResponseDto
 {
-    public mixed $value;
+    use HasTimestamps;
 
     public int $id;
 
+    public string $key;
+
     public CustomFieldDtoTypeEnum $type;
+
+    public string $modelType;
+
+    public ?int $customFieldGroupId;
+
+    public mixed $params;
 
     public bool $isPrivate;
 
@@ -24,38 +40,39 @@ class CustomFieldDto extends ApiResponseDto
      */
     public Collection $translations;
 
-    public ?CustomFieldGroupDto $group;
-
-    public ?object $params;
-
     protected function getDefaultDataStructureToProperties(): array
     {
-        return [
-            'value'  => 'value',
-            'config' => [
-                'id'           => 'id',
-                'type'         => fn ($type) => $this->type = CustomFieldDtoTypeEnum::from($type),
-                'private'      => 'isPrivate',
-                'translations' => fn ($translations) => $this->initTranslations($translations),
-                'group'        => fn ($group) => $this->group = is_null($group) ? null : new CustomFieldGroupDto($group),
-                'params'       => 'params',
-            ],
+        $structure = [
+            'id'                    => 'id',
+            'key'                   => 'key',
+            'type'                  => fn ($type) => $this->type = CustomFieldDtoTypeEnum::tryFrom($type),
+            'model_type'            => 'modelType',
+            'custom_field_group_id' => 'customFieldGroupId',
+            'params'                => 'params',
+            'private'               => 'isPrivate',
+            // Fractal wraps every include (even nested ones) in its own `{"data": [...]}` envelope.
+            'translations' => fn ($translations) => $this->initTranslations($translations->data ?? []),
         ];
+
+        return array_merge($structure, $this->getTimestampsAsDataStructureToProperties());
     }
 
     protected function validationRulesForProperties(): array
     {
         $types = collect(CustomFieldDtoTypeEnum::cases())->pluck('value');
 
-        return [
-            'value'               => ['present'],
-            'config.id'           => ['required', 'int'],
-            'config.type'         => ['required', Rule::in($types)],
-            'config.private'      => ['required', 'boolean'],
-            'config.translations' => ['required', 'array'],
-            'config.group'        => ['nullable', 'array'],
-            'config.params'       => ['present'],
+        $rules = [
+            'id'                    => ['required', 'integer'],
+            'key'                   => ['required', 'string'],
+            'type'                  => ['required', Rule::in($types)],
+            'model_type'            => ['required', 'string'],
+            'custom_field_group_id' => ['present', 'nullable', 'integer'],
+            'params'                => ['present', 'nullable'],
+            'private'               => ['required', 'boolean'],
+            'translations'          => ['required', 'array'],
         ];
+
+        return array_merge($rules, $this->getTimestampsValidationRules());
     }
 
     protected static function getDefaultNotConstructableException(
