+++ src/Exceptions/Dtos/CustomField/CustomFieldModelValueDtoNotConstructableException.php
@@ -4,4 +4,4 @@ namespace OrigamiMp\OrigamiApiSdk\Exceptions\Dtos\CustomField;
 
 use OrigamiMp\OrigamiApiSdk\Exceptions\Dtos\ApiResponseDtoNotConstructableException;
 
-class CustomFieldDefinitionListDtoNotConstructableException extends ApiResponseDtoNotConstructableException {}
+class CustomFieldModelValueDtoNotConstructableException extends ApiResponseDtoNotConstructableException {}
