+++ src/Repositories/Api/Data/CustomField/OrigamiCustomFieldDataApiRepository.php
@@ -2,7 +2,7 @@
 
 namespace OrigamiMp\OrigamiApiSdk\Repositories\Api\Data\CustomField;
 
-use OrigamiMp\OrigamiApiSdk\Dtos\CustomField\CustomFieldDefinitionListDto;
+use OrigamiMp\OrigamiApiSdk\Dtos\CustomField\CustomFieldListDto;
 use OrigamiMp\OrigamiApiSdk\Dtos\CustomField\CustomFieldValueListDto;
 use OrigamiMp\OrigamiApiSdk\ParamBags\Data\CustomField\ListCustomFieldRequestParamBag;
 use OrigamiMp\OrigamiApiSdk\ParamBags\Data\CustomField\ListCustomFieldValueRequestParamBag;
@@ -10,12 +10,12 @@ use OrigamiMp\OrigamiApiSdk\Repositories\Api\Data\OrigamiDataApiRepository;
 
 class OrigamiCustomFieldDataApiRepository extends OrigamiDataApiRepository
 {
-    public function list(ListCustomFieldRequestParamBag $paramBag): CustomFieldDefinitionListDto
+    public function list(ListCustomFieldRequestParamBag $paramBag): CustomFieldListDto
     {
         $response = $this->restClient->get('custom_fields', $paramBag);
         $responseContent = json_decode($response->getBody()->getContents());
 
-        return new CustomFieldDefinitionListDto($responseContent);
+        return new CustomFieldListDto($responseContent);
     }
 
     /**
